refactor: isPlaying 상태 분리 및 usePlayingStore 생성
- useChannelStore에서 isPlaying 관련 상태와 로직을 분리
- isPlaying 관리를 위한 usePlayingStore 신규 생성
- useChannelStore는 관심 채널, 이전 채널, 라디오 채널 리스트 등만 관리하도록 변경
- channel-storage로 persist 네이밍 명확화
